Group eligibility screener controllers into a folder
Makes is easier to add new pages in the future.
