+++ config/routes.rb
@@ -45,21 +45,21 @@ Rails.application.routes.draw do
     get "/start", to: "pages#start"
     get "/users/registrations/exists", to: "users/existing_registration#new"
     post "/users/registrations/exists", to: "users/existing_registration#create"
-    get "/referral-type", to: "referral_type#new"
-    post "/referral-type", to: "referral_type#create"
-    get "/have-you-complained", to: "have_complained#new"
-    post "/have-you-complained", to: "have_complained#create"
+    get "/referral-type", to: "eligibility_screener/referral_type#new"
+    post "/referral-type", to: "eligibility_screener/referral_type#create"
+    get "/have-you-complained", to: "eligibility_screener/have_complained#new"
+    post "/have-you-complained", to: "eligibility_screener/have_complained#create"
     get "/no-complaint", to: "pages#no_complaint"
-    get "/is-a-teacher", to: "is_teacher#new"
-    post "/is-a-teacher", to: "is_teacher#create"
-    get "/unsupervised-teaching", to: "unsupervised_teaching#new"
-    post "/unsupervised-teaching", to: "unsupervised_teaching#create"
+    get "/is-a-teacher", to: "eligibility_screener/is_teacher#new"
+    post "/is-a-teacher", to: "eligibility_screener/is_teacher#create"
+    get "/unsupervised-teaching", to: "eligibility_screener/unsupervised_teaching#new"
+    post "/unsupervised-teaching", to: "eligibility_screener/unsupervised_teaching#create"
     get "/no-jurisdiction-unsupervised", to: "pages#no_jurisdiction_unsupervised"
-    get "/teaching-in-england", to: "teaching_in_england#new"
-    post "/teaching-in-england", to: "teaching_in_england#create"
+    get "/teaching-in-england", to: "eligibility_screener/teaching_in_england#new"
+    post "/teaching-in-england", to: "eligibility_screener/teaching_in_england#create"
     get "/no-jurisdiction", to: "pages#no_jurisdiction"
-    get "/serious-misconduct", to: "serious_misconduct#new"
-    post "/serious-misconduct", to: "serious_misconduct#create"
+    get "/serious-misconduct", to: "eligibility_screener/serious_misconduct#new"
+    post "/serious-misconduct", to: "eligibility_screener/serious_misconduct#create"
     get "/not-serious-misconduct", to: "pages#not_serious_misconduct"
     get "/you-should-know", to: "pages#you_should_know"
     get "/complete", to: "pages#complete"
