+++ spec/forms/eligibility_screener/is_teacher_form_spec.rb
@@ -1,27 +1,25 @@
 require "rails_helper"
 
-RSpec.describe UnsupervisedTeachingForm, type: :model do
+RSpec.describe EligibilityScreener::IsTeacherForm, type: :model do
   describe "validations" do
     subject(:form) { described_class.new }
 
     it { is_expected.to validate_presence_of(:eligibility_check) }
 
-    specify do
-      expect(form).to validate_inclusion_of(:unsupervised_teaching).in_array(%w[yes no not_sure])
-    end
+    specify { expect(form).to validate_inclusion_of(:is_teacher).in_array(%w[yes no not_sure]) }
   end
 
   describe "#valid?" do
     subject(:valid) { form.valid? }
 
     let(:eligibility_check) { EligibilityCheck.new }
-    let(:form) { described_class.new(eligibility_check:, unsupervised_teaching:) }
-    let(:unsupervised_teaching) { "yes" }
+    let(:form) { described_class.new(eligibility_check:, is_teacher:) }
+    let(:is_teacher) { "yes" }
 
     it { is_expected.to be_truthy }
 
-    context "when unsupervised_teaching is blank" do
-      let(:unsupervised_teaching) { "" }
+    context "when is_teacher is blank" do
+      let(:is_teacher) { "" }
 
       it { is_expected.to be_falsy }
     end
@@ -31,11 +29,11 @@ RSpec.describe UnsupervisedTeachingForm, type: :model do
     subject(:save) { form.save }
 
     let(:eligibility_check) { EligibilityCheck.new }
-    let(:form) { described_class.new(eligibility_check:, unsupervised_teaching: "yes") }
+    let(:form) { described_class.new(eligibility_check:, is_teacher: "yes") }
 
     it "saves the eligibility check" do
       save
-      expect(eligibility_check.unsupervised_teaching).to be_truthy
+      expect(eligibility_check.is_teacher).to be_truthy
     end
   end
 end
