+++ spec/forms/eligibility_screener/unsupervised_teaching_form_spec.rb
@@ -1,13 +1,13 @@
 require "rails_helper"
 
-RSpec.describe SeriousMisconductForm, type: :model do
+RSpec.describe EligibilityScreener::UnsupervisedTeachingForm, type: :model do
   describe "validations" do
     subject(:form) { described_class.new }
 
     it { is_expected.to validate_presence_of(:eligibility_check) }
 
     specify do
-      expect(form).to validate_inclusion_of(:serious_misconduct).in_array(%w[yes no not_sure])
+      expect(form).to validate_inclusion_of(:unsupervised_teaching).in_array(%w[yes no not_sure])
     end
   end
 
@@ -15,13 +15,13 @@ RSpec.describe SeriousMisconductForm, type: :model do
     subject(:valid) { form.valid? }
 
     let(:eligibility_check) { EligibilityCheck.new }
-    let(:form) { described_class.new(eligibility_check:, serious_misconduct:) }
-    let(:serious_misconduct) { "yes" }
+    let(:form) { described_class.new(eligibility_check:, unsupervised_teaching:) }
+    let(:unsupervised_teaching) { "yes" }
 
     it { is_expected.to be_truthy }
 
-    context "when serious_misconduct is blank" do
-      let(:serious_misconduct) { "" }
+    context "when unsupervised_teaching is blank" do
+      let(:unsupervised_teaching) { "" }
 
       it { is_expected.to be_falsy }
     end
@@ -31,11 +31,11 @@ RSpec.describe SeriousMisconductForm, type: :model do
     subject(:save) { form.save }
 
     let(:eligibility_check) { EligibilityCheck.new }
-    let(:form) { described_class.new(eligibility_check:, serious_misconduct: "yes") }
+    let(:form) { described_class.new(eligibility_check:, unsupervised_teaching: "yes") }
 
     it "saves the eligibility check" do
       save
-      expect(eligibility_check.serious_misconduct).to be_truthy
+      expect(eligibility_check.unsupervised_teaching).to be_truthy
     end
   end
 end
